Moved CtfFlag to Schema folder
